Move get_device and get_devices to DeviceManager
